+++ bin/miden-cli/src/errors.rs
@@ -1,3 +1,5 @@
+#![allow(unused_assignments)]
+
 use std::error::Error;
 
 use miden_client::account::{AccountId, AddressError};
@@ -38,7 +40,6 @@ pub enum CliError {
     #[diagnostic(code(cli::client_error))]
     Client {
         #[source]
-        #[allow(unused_assignments)]
         error: ClientError,
         #[help]
         help: Option<String>,